requirements.txt
# altair==5.0.1
# attrs==23.1.0
# blinker==1.6.2
# cachetools==5.3.1
# certifi==2023.5.7
# charset-normalizer==3.1.0
# click==8.1.3
# contourpy==1.1.0
# cycler==0.11.0
# decorator==5.1.1
# filelock==3.12.2
# fonttools==4.40.0
# gitdb==4.0.10
# GitPython==3.1.31
# h11==0.14.0
# idna==3.4
# importlib-metadata==6.7.0
# Jinja2==3.1.2
# jsonschema==4.17.3
# kiwisolver==1.4.4
# markdown-it-py==3.0.0
# MarkupSafe==2.1.3
# matplotlib==3.7.1
# mdurl==0.1.2
# mpmath==1.3.0
# networkx==3.1
# numpy==1.25.0
# opencv-python==4.7.0.72
# packaging==23.1
# pandas==2.0.2
# Pillow==9.5.0
# protobuf==4.23.3
# psutil==5.9.5
# pyarrow==12.0.1
# pydeck==0.8.1b0
# Pygments==2.15.1
# Pympler==1.0.1
# pyparsing==3.1.0
# pyrsistent==0.19.3
# python-dateutil==2.8.2
# pytz==2023.3
# pytz-deprecation-shim==0.1.0.post0
# PyYAML==6.0
# requests==2.31.0
# rich==13.4.2
# scipy==1.10.1
# seaborn==0.12.2
# six==1.16.0
# smmap==5.0.0
# streamlit==1.23.1
# sympy==1.12
# tenacity==8.2.2
# toml==0.10.2
# toolz==0.12.0
# torch==2.0.1
# torchvision==0.15.2
# tornado==6.3.2
# tqdm==4.65.0
# typing_extensions==4.6.3
# tzdata==2023.3
# tzlocal==4.3.1
# ultralytics==8.0.121
# urllib3==2.0.3
# uvicorn==0.22.0
# validators==0.20.0
# zipp==3.15.0
# # streamlit==1.23.1
# # ultralytics==8.0.121
# # uvicorn==0.22.0
streamlit
pillow
opencv-python
numpy
matplotlib
ultralytics